qmake: fix return type of toLongLong()
Fixes: QTBUG-71886

+++ qmake/library/proitems.h
@@ -133,7 +133,7 @@ public:
     bool contains(const QString &s, Qt::CaseSensitivity cs = Qt::CaseSensitive) const { return indexOf(s, 0, cs) >= 0; }
     bool contains(const char *s, Qt::CaseSensitivity cs = Qt::CaseSensitive) const { return indexOf(QLatin1String(s), 0, cs) >= 0; }
     bool contains(QChar c, Qt::CaseSensitivity cs = Qt::CaseSensitive) const { return indexOf(c, 0, cs) >= 0; }
-    int toLongLong(bool *ok = nullptr, int base = 10) const { return toQStringRef().toLongLong(ok, base); }
+    qlonglong toLongLong(bool *ok = nullptr, int base = 10) const { return toQStringRef().toLongLong(ok, base); }
     int toInt(bool *ok = nullptr, int base = 10) const { return toQStringRef().toInt(ok, base); }
     short toShort(bool *ok = nullptr, int base = 10) const { return toQStringRef().toShort(ok, base); }
